Track approvals slightly differently
We now track both approved and nont yet approved checkers where
we tracked only the approved ones before.  I'm torn over this
change, but it's a step towards a more efficent solution,
which we may or may not require.  I'll make this change now
in the hopes that it simplifies something later.